requirements.txt
Django==5.0.6
djangorestframework==3.16.1
firebase-admin==7.1.0